Add ability to open cookie database file with exclusive flag.
This CL adds the ability to specify that the sqlite file containing
cookies be opened with the `exclusive` uri parameter. On Windows,
this instructs sqlite to open it without the FILE_SHARE_READ and
FILE_SHARE_WRITE sharing dispositions, meaning that handles can
no longer be obtained to the file while the database is open.

The flag is wired up through the net/extras/sqlite layer and into
the network service, where it is exposed as an optional parameter
on network_context.

In //chrome, the flag is set on Windows for profile based cookie
stores, based on the feature LockProfileCookieDatabase, which is
disabled by default.

Tests are added at each layer to verify the behavior:
 - sql_unittests ->
    that the new parameter enable_exclusive_database_file_lock is
    respected.
 - net_unittests ->
    that the new parameter to SQLitePersistentCookieStore is
    respected.
 - browser_tests ->
    that the feature LockProfileCookieDatabase successfully
    toggles the feature for the profile cookie store.

BUG=1429117

Change-Id: I2beaa0eab330535eecf733ecc8fd6a93f748227a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4378424
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Austin Sullivan <asully@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1125803}
NOKEYCHECK=True
GitOrigin-RevId: 4e37d8b3fb5edac4e61bc568a26212b299fe2676