[ios] Fix assertions comparing pointer for equality
WebState::GetStableIdentifier() returns a NSString* and DCHECK_EQ()
macro uses ==, thus the assertions were comparing the pointer value
instead of the string.

As WebState::GetStableIdentifier() is initialized from a -copy of
the identifier in CRWSessionStorage, the pointer happened to be
equal most of the time, but this assumption may be incorrect in
certain cases.

Since WebState and CRWSessionStorage exposes another identifier that
can be compared using == (as they are SessionID), use this instead.
This has the added benefit that DCHECK_EQ() will correctly display
the value if the assertion fails.

Bug: 1428854, 1383087, 1430399
Change-Id: I1c77cd2ece26ac5a668fdd0c8880256d78ddbd32
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4396360
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Petro Akzhygitov <pakzhygitov@google.com>
Commit-Queue: Petro Akzhygitov <pakzhygitov@google.com>
Cr-Commit-Position: refs/heads/main@{#1125993}
NOKEYCHECK=True
GitOrigin-RevId: 05ec4a78872673431ad77cc1080ffc0c9b6083d8