[M113][iOS] Fix crash when adding to reading list from incognito
The new snackbar shown when an item is added to reading list shows the
primary user's email if the sync is on. We retrieve this from the
IdentityManager, but the latter is null if created from incognito
browser state.

As for bookmarks, reading list items of the original primary account is
shown in incognito mode. Therefore it seems natural to show the
account's email in incognito mode as well when adding a reading list
entry.

This CL ensures that the IdentityManager used is retrieved using the
original browser state, and adds an EGTest for adding an item in incognito mode.

Bug: 1428055
Change-Id: I2e7ba50cc456d01f14c860361a1e79466bef8460
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4386662
Commit-Queue: Menghan Yang <myuu@google.com>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1126470}
NOKEYCHECK=True
GitOrigin-RevId: 0e714ac98ea7d71e5fb215c1acf38acf78b74953