[iOS] Add standard sign-in promo in reading list
This CL adds the standard sign-in promo view to the reading list UI.
The sign-in promo should be shown only when the reading list account
storage is enabled.

This promo will be adapted in a following CL to show a loader during
the reading list initial sync after sign-in, and to allow sign-in
without full sync.

Screenshots:
https://drive.google.com/drive/folders/1ukMQXxalaFquTD7IglrmB9OMtIt1wUJ2?usp=sharing

* crrev.com/c/4342389 [iOS] Add sign-in access point for reading list
=>crrev.com/c/4359675 [iOS] Add sign-in promo in reading list

Bug: 1420194
Change-Id: If58e16d3d85d9dc80a48c55565eb162ae957e5a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4359675
Commit-Queue: Menghan Yang <myuu@google.com>
Reviewed-by: Petro Akzhygitov <pakzhygitov@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1126670}
NOKEYCHECK=True
GitOrigin-RevId: 813c03574b3b656341592e212d2db0c77196cc21