[iOS] Action to add and edit bookmark
The action name indicated that it was here to track "add bookmark",
but actually it was called on the button that add or edit a bookmark.
So I renamed it to be clearer.

I also change the owner of some bookmark related action. Previous
owner has not been part of chrome for 7 years.

Bug: 1430394
Change-Id: I6509912c81bd00125a6c11e0af44a66ded911fbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4395792
Reviewed-by: Daniel Classon <dclasson@google.com>
Commit-Queue: Arthur Milchior <arthurmilchior@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Petro Akzhygitov <pakzhygitov@google.com>
Cr-Commit-Position: refs/heads/main@{#1126825}
NOKEYCHECK=True
GitOrigin-RevId: be1e467311d87fcdc3635163cdf0fc2c3cc1fc9f