[ios] Dont verify position of highlighted destinations in egtest
Many test runs failed in builders because the position of the
destination wouldn't be the expected one.

The position of the destination can move because of factors beyond
the test case. Until we find a more reliable approach for verifying
the position, we should not verify it.

Change-Id: I19d279a4047fcb45345b2dbcaa55efbc5316c50a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4401463
Owners-Override: Louis Romero <lpromero@google.com>
Reviewed-by: Louis Romero <lpromero@google.com>
Commit-Queue: Vincent Boisselle <vincb@google.com>
Cr-Commit-Position: refs/heads/main@{#1126852}
NOKEYCHECK=True
GitOrigin-RevId: ea9baf81196ab3a408702c7753c2715c6eab6686