[ios] Remove NTPCoordinator -start call from -displayWebState:
This -start call catches all instances where the BVC's presenting
view is the NTP and it has yet to be started. Namely:
1) Whenever inserting a tab showing the NTP from a batch tab insertion
2) Whenever re-activating the Browser (i.e. clear browsing data) while
currently showing the NTP

In order to handle these situations, -start is added in the relevant
locations. In instances where the WebState showing the NTP is not the
active WebState, -startNTPIfNeededForActiveWebState: is added so that
when that WebState is activated, it is started. It could also just
be started in the relevant location, but would require complexity
in NTPCoordinator's -start to get the right WebState and would make any
WebState switch seem internally like a switch away from the NTP.

This change also moves the -stop call in uninstallDependencyForWebState:
to didDetachWebState: and didReplaceWebState: in TabEventsMediator,
which are the caller sources of uninstallDependencyForWebState:.

Bug: 1406940, 1411808
Change-Id: I49064cae1bab5ddf73ea0d1933c0eeeb2dee6a1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4375783
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1127228}
NOKEYCHECK=True
GitOrigin-RevId: f400556cfe5bee09aaa53a9433753136f2d238d1