[ios] Make SnapshotTabHelper own the snapshot identifier
In preparation of removing GetStableIdentifier(), move the ownership
of the snapshot identifier to SnapshotTabHelper. This centralize the
call to GetStableIdentifier() in a single location, which will allow
to convert it without breakage.

Modernize SnapshotGenerator to use ivar directly instead of private
properties (and to use the ivar to access the delegate and the cache
instances as this simplify the code generated).

Bug: 1430399
Change-Id: I917728c59f5fc434bc64ba144a0b6cff9f953b31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4395999
Reviewed-by: Federica Germinario <fedegermi@google.com>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128609}
NOKEYCHECK=True
GitOrigin-RevId: 472dc254796cb9e1770d8f249e3143842603a72d