Support having Inactive Tabs settings be presented from Inactive Tabs
Prior to this CL, when opening the Inactive Tabs settings from the
Inactive Tabs grid, the InactiveTabsSettingsMediator got the inactive
browser, which broke the migration code.
With this CL, we support launching the settings with any of the active
or inactive browser. The mediator then retrieves the associated pair of
browsers and migrate accordingly.

Bug: 1429294
Change-Id: Iaf160d982c985f65dcfbd500957dbf051df96825
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4395998
Reviewed-by: Aliona Dangla <alionadangla@chromium.org>
Commit-Queue: Louis Romero <lpromero@google.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Auto-Submit: Louis Romero <lpromero@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128657}
NOKEYCHECK=True
GitOrigin-RevId: ca073956e7c175370a35c2c75655171dc4ecc867