[IOS][Omnibox] Fix test scrolling dismisses keyboard
The test is failing mainly on iPhone SE simulator.
It is failing to capture the top row suggestion.
The reason for that is that when we swipe up, the row is hidden (small
screen).
One way to solve this is to swipe down instead so that we'll always
have the top row displayed.

Bug: 1430354
Change-Id: Ia69edcee917ef51b2e9bae180e522e7d84d20343
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4414736
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Ameur Hosni <ameurhosni@google.com>
Reviewed-by: Christian Xu <christianxu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1128686}
NOKEYCHECK=True
GitOrigin-RevId: b07c778f2dfea3442a259aa269839c8009acd32e