Do not reuse TranslateInfobarDelegate on new translate flow
It is possible to trigger a new translate flow while the
infobar is displayed.
In that case, updating the delegate is not enough as it will
not update the UI accordingly.
Remove the old infobar and trigger a new one to keep it in
a consistent state with the page.

Bug: 1421728
Change-Id: Ia32b437f40cf17ca7a856ed19b588b389e3d38fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4403592
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Anthony Cui <cuianthony@chromium.org>
Auto-Submit: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1129234}
NOKEYCHECK=True
GitOrigin-RevId: ae50269a807d3afc9a4107475dbfc97ff9e5fb7a