[ios][BVC] Move addURLsToReadingList out of BVC
This CL moves addURLsToReadingList and addURLToReadingList inside
ReadingListBrowserAgent.

As a result, BVC does not need anymore the following objects:
 - SnackbarCommandsHandler
 - ToolbarCommandsHandler
 - ReadingModel
 - IdentityManager
On the other hand, knowledge of ReadingListBrowserAgent is now required.

Note that this is only a first step, the final objective is to
remove addToReadingList from BrowserCommands and from BVC (once done
BVC will not need anymore ReadingListCoordinator). This will be handled
in separate CLs since addToReadingList is used in many parts of the
code (also in ios_internal).

Bug: 1272540
Change-Id: I78615ec8993e7d08eab38fa9c8f1ce94590adff3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4373297
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Federica Germinario <fedegermi@google.com>
Cr-Commit-Position: refs/heads/main@{#1129739}
NOKEYCHECK=True
GitOrigin-RevId: 855d063c1fedbd07b5efbb11e32bba53f6af61d3