[ios] Add a CrossfadeLabel subclass of UILabel
Some attributes of UILabel, such as textColor or text attributes such as
strikethrough, are not able to be animated via UIKit's normal animation
methods. This class enables a label to animate those attributes by
creating a copy of itself with the desired changes, then overlaying that
copy over itself in its superview, and crossfading to display the copy.
When the crossfade is complete, it copies some changed attributes from
the overlay back to itself, shows itself again, and removes the copy.

This functionality will be used by the Set Up List feature. An example
of its usage can be seen here:
https://drive.google.com/file/d/1F7A9iAY8nVDzXZCwAgnnPbMVwckmL605/view?usp=sharing

Bug: 1427439
Change-Id: I0bb6416781777c118c2388ccc96375c24b173b43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4404705
Commit-Queue: Scott Yoder <scottyoder@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1129883}
NOKEYCHECK=True
GitOrigin-RevId: 5c9b5d6cbcb04238ec5c02850efdb536b793885a