[iOS] Extract visibility changes from NTP start/stop
Remove NTP visibility changes from the coordinator's start/stop
functions. Starting the NTPCoordinator should not necessarily mean
that the NTP has become visible, ie. when creating an NTP in an
inactive web state. Therefore, these calls should be separate.

This fixes an issue where NTPDidChangeVisibility:YES would be
triggered on an already visible NTP when switching from one NTP
to another.

This is a step in the direction of having more meaningful and
intentional start/stop calls, as it allows the coordinator lifecycle
to be controlled without necessarily updating the NTP visibility.

By separating navigation calls between the NTPTabHelper and the TabEventsMediator, we'll be able to distinguish NTP visibility
changes between web states or within the same web state. This will
be done in a follow-up CL and will help with things like saved scroll
position, as it will be clearer when a user navigated back to an NTP expecting to retain their position.

Small refactoring of NTP visibility function, removing unnecessary properties like `sceneInForeground` and `viewPresented`.

Change-Id: I5b71b4a945234e2c7d02c5c7eb7960c39c42baae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4359775
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Adam Arcaro <adamta@google.com>
Cr-Commit-Position: refs/heads/main@{#1130015}
NOKEYCHECK=True
GitOrigin-RevId: 8d0480e1672dfd352d7c87e064e590d1935dd986