[iOS] Add DetailsContext to PasswordDetails
In the scope of the Password Checkup feature, the presence of some
elements in the PasswordDetailsTableViewController depends on its entry
point. This CL adds the DetailsContext enum which allow the
PasswordDetailsTableViewController to access what the entry point was
and adapt its content accordingly.

Bug: 1406540
Change-Id: I6731fe439b0fb17d00add1ccf0504318bc4314dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4415224
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Noémie St-Onge <noemies@google.com>
Cr-Commit-Position: refs/heads/main@{#1130070}
NOKEYCHECK=True
GitOrigin-RevId: 92f5c499240f17510fa76de0123a164a9f22f263