Fix crash when PasswordDetails is opened and all open tabs are incognito
This fixes a crash introduced in M113 (crrev.com/c/4290479). Before that
CL, the problematic code was only hit if EnablePasswordsAccountStorage
was enabled and the user tried to move a local password to the account.
Now it's hit regardless of any flags, as long as one follows the
repro instructions in the title of this CL.

The root cause is settings coordinators receive a non-incognito Browser
even when they are started from incognito. Looking for a tab in the
non-incognito browser will fail if only incognito tabs are open.

The adopted fix is to look for tabs also in other browsers, using
BrowserList. A better fix would be to refactor some of the code
//components/password_manager/ to avoid unnecessarily depending on
tabs. But that's a much larger effort and riskier to merge back.

Fixed: 1431975
Change-Id: I8126352f7a9a0ef5b7fd38ed9e8be8a5e7ba6ddb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4418726
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Auto-Submit: Victor Vianna <victorvianna@google.com>
Quick-Run: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/main@{#1130583}
NOKEYCHECK=True
GitOrigin-RevId: 49166a432e41ff46a63add3f5cf6ff2851a0e21a