[iOS] Close all from regular tab grid close inactive's tabs too
In this CL, inactive tabs are added in the tab restore service after
the active ones and are restore before the active ones to keep the
correct order. It also creates the InactiveTabsCoordinator when the app
is launched, and not any more just when the UI is first shown because
the InactiveTabsMediator need to be always alive.

See video:
https://drive.google.com/file/d/12HTjCe5SF4OwyUOPhiVsTsUqoM4aQzA1/view?usp=share_link

Fixed: 1431152
Change-Id: Ia75ba6396a8c80f599093f0bfee2b1ba28a64fa9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4403962
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Aliona Dangla <alionadangla@chromium.org>
Reviewed-by: Louis Romero <lpromero@google.com>
Cr-Commit-Position: refs/heads/main@{#1130610}
NOKEYCHECK=True
GitOrigin-RevId: f1812fae8a2d7cbd3d9e35a231006468b7d4d5dc