[ios] Stop saving/restoring PageDisplayState
This was used to serialize the zoom level and scroll position of a
page when saving the session, however, this is only used when the
feature kRestoreSessionFromCache is disabled or when running on a
version of iOS older than 15.0.

kRestoreSessionFromCache is launched at 100% since M-100 and the
deployment target has been bumped in https://crrev.com/c/4381944
to iOS 15.0+, this means that this is only used when the WebKit
sessionState cannot be saved or is too large to be saved. It is
not deemed worth keeping the code for those unlikely scenarios.

This is part of a multi-CL change:
 => https://crrev.com/c/4420362
 -  https://crrev.com/c/4427830
 -  https://crrev.com/c/4427793
 -  https://crrev.com/c/4427972
 -  https://crrev.com/c/4427852

Bug: none
Change-Id: I3938fb78a681aa456712112610b63c1011e08fd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4420362
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1131229}
NOKEYCHECK=True
GitOrigin-RevId: cf1a2d89ff1a290a9ed23f0b9ff287b2f34904bd