[ios] Switch "Share Chrome" to use iOS app store URL w/ param
We were unable to use the cross-platform URL w/ a param, but we can use
the direct iOS app store URL with a param in order to allow us to track
how many installs originate from this feature.

Note: the URL parameter is static. It does not, for example, allow
tracking a user in any way. It only allows tracking usage of the
feature.

Bug: 1340308
Change-Id: Ia4a708726c13821c9fbcc69e67ee9db968875e81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4425231
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1131233}
NOKEYCHECK=True
GitOrigin-RevId: ef9c3b737dd44be61c80eceaadcdcb299f6542fc