[iOS] Fix PasswordCheckupViewController cells height not adjusting
Sometimes, the subline of some of the cells in
PasswordCheckupViewController goes from 1 line to 2. When it happened,
the height of the cells weren't adjusting their height correctly. This
CL fixes the issue by updating the cells content inside a
`performBatchUpdates` block.

Bug: 1406540
Change-Id: I8d6d6a409f9dd6a21070cda49252c35f5978c000
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4433610
Reviewed-by: Ernesto Izquierdo Clua <eic@google.com>
Commit-Queue: Noémie St-Onge <noemies@google.com>
Cr-Commit-Position: refs/heads/main@{#1131244}
NOKEYCHECK=True
GitOrigin-RevId: cb80cf91785fafc73d8bc5a3363661bb3444bd51