[iOS] Show undo action in item added to reading list snackbar
Show the Undo button in the snackbar which is shown when one or several
items are added to the reading list, which removes the newly added
entries.

This should be visible only if the reading list account storage and
related UI are enabled, and if the new snackbar with account email is
being used.

The undo action is not a perfect restore of the previous state, and is
an actual removal of the item. It means that:
- It will remove the item from both storages if the account storage is
enabled. So if an entry was local, then added by the user to the account
storage, and then the adding is undone, the entry will be removed from
the local and the remote storage.
- If the user tries to re-add an existing entry (no-op add), then taps
"undo", the existing entry will be removed.

Screenshot:
https://drive.google.com/file/d/1BgWoVvlg8a52TAcDJ-qdmGw22iL-Ce6U/view?usp=share_link

Fixed: 1420287
Change-Id: I1b889ed2268ac73362a552cb5a69afd5657a2738
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4374539
Commit-Queue: Menghan Yang <myuu@google.com>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1131331}
NOKEYCHECK=True
GitOrigin-RevId: 6935b7862bae11ce9d6a5429d031043dfe92462a