[iOS] Fix PasswordCheckupUtils::GetWarningOfHighestPriority
When an insecure password was both muted and weak/reused,
GetWarningOfHighestPriority was prioritizing the fact that it was muted,
which is not the desired behaviour. As stated in the method comment of
GetWarningOfHighestPriority, muted password warnings have a lower
priority than reused and weak passwords. This CL fixes the logic issue.

Bug: 1406540
Change-Id: I5052e2227d2827ad97e7fcc481e22ef1c566d181
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4430453
Reviewed-by: Ernesto Izquierdo Clua <eic@google.com>
Commit-Queue: Noémie St-Onge <noemies@google.com>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1131485}
NOKEYCHECK=True
GitOrigin-RevId: 70d2c26153e8c439e3328e6af28847afec17bb15