[iOS] Return not Start surface if NTP is not visible
When toggling sign-in with an NTP in another tab while the promo is
visible, the promo checks if the NTP is start surface. Since
background NTPs don't have web states, this crashes the app.

The NTP is never a start surface if it is in a background tab, so
we return NO if not `visible`.

Fixed: 1425382
Change-Id: Ibcb0a6fe71b9fb9e53e2f72837c456ef3c9869ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4441546
Auto-Submit: Adam Arcaro <adamta@google.com>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Adam Arcaro <adamta@google.com>
Cr-Commit-Position: refs/heads/main@{#1131966}
NOKEYCHECK=True
GitOrigin-RevId: 9662e7a7c31ec6078379e4748a2cb6b575ecc911