Remove GetMainWebFrameId and GetWebFrameWithId
This util only supports the page content world which is less widely used
now than it was when the utility was created. Additionally, it can hide
this fact from the caller and lead to bugs where the expected frames are
not returned (because the associated features live in an isolated
world).

Bug: 1427327
Change-Id: I2c783f174bacbfeeaf83d063832bf2cb71d82bc4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4442406
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1132304}
NOKEYCHECK=True
GitOrigin-RevId: 070c52e2900f56e6cce792062ef40200936db9cb