[iOS] Make bookmark account storage promo reshow after sign-in & out
The new account storage promo for bookmarks should be hidden on user
sign-in and reshow after sign-out. So it is needed to capture and handle
primary account change for the signin consent level (instead of sync).

Bug: 1432611
Change-Id: I67087d0423b63189d23c2fba53c00af18d9d2d27
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4439614
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Menghan Yang <myuu@google.com>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1132411}
NOKEYCHECK=True
GitOrigin-RevId: 433b8c4e74b8ebf1269e0df60509987aaea45b8d