[iOS] On/Off String on Push Notification Settings Page
On the settings page, the push notification setting tile should display
either the On or Off string to the left of the chevron if all of the
push notification enabled feature's push notification permissions are
enabled or all disabled. The On or Off string should also appear on each
of the push notification enabled feature's tile given the same criteria.

Bug: 1431473
Change-Id: I15c24db46a25c27ad97904a146825ea26012f9ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4403768
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Daniel White <danieltwhite@google.com>
Cr-Commit-Position: refs/heads/main@{#1132480}
NOKEYCHECK=True
GitOrigin-RevId: 70ff2d3d8e5abfc779605a3fcc96421d9ae7a6f7