[iOS] Do not have a moved tab in recently closed list
As recently closed tab are added by a web state list observer, it adds
any web state that have been "detached". When a tab is moved, it is not
closed, so it should not appear in recently close tab. This CL moves
the logic to save a tab in recently close in `WillCloseWebStateAt` call
instead of `WillDetachWebStateAt`.

See video:
https://drive.google.com/file/d/10tNhhJgPhbxEe_IlS0AwtKTb2s-UuPWw/view?usp=share_link

Bug: 1414048
Fixed: 1433323
Change-Id: I1bfdbc04625c54db121d0049fc35e5739e632790
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4427979
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Louis Romero <lpromero@google.com>
Commit-Queue: Aliona Dangla <alionadangla@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1132542}
NOKEYCHECK=True
GitOrigin-RevId: cc20317a4b653d241fd625c4a0ca6162df061453