[iOS] Handle case where there is only 1 password left in a reused group
When passwords are identified as reused, all the credentials with the
same password are grouped together (reused group). If all the
passwords in that group are deleted or modified except for one, that
one password left is still marked as reused (which is not accurate
anymore). This CL brings a temporary solution to filter out a reused
password if it's the only one left in its group. A better solution
will be implemented in the future (see crbug.com/1434343).

Fixed: 1434073
Change-Id: I58959fe7e173d3a00f4c5e6f65d88e808986e9a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4436611
Reviewed-by: Ernesto Izquierdo Clua <eic@google.com>
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Commit-Queue: Noémie St-Onge <noemies@google.com>
Cr-Commit-Position: refs/heads/main@{#1133406}
NOKEYCHECK=True
GitOrigin-RevId: be652758110de866e92fb9b756091194e1323f7e