[iOS] Make Reading list account storage promo reshow after sign-in & out
The new account storage promo for reading list should be hidden on user
sign-in and reshow after sign-out. So it is needed to capture and handle
primary account change for the sign-in consent level (instead of sync).

This CL updates the consent level and adds related EGtest.

Fixed: 1432607
Change-Id: I0f4e125eae4c69b4bda310b400c9a78e057f19f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4446362
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Menghan Yang <myuu@google.com>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1133662}
NOKEYCHECK=True
GitOrigin-RevId: eda4b71f827a6bdcae991baf3816075dd077e666