[iOS] Fixed Cold Start Push Notification Redirection Issue
When the user received a push notification from price tracking and
Chrome was not running on the device, the user was not redirected to
the product's URL on interacting with the push notification.

This was due to the fact that the UNNotificationCenter Delegate
initialization occurred in the firstSceneWillEnterForeground step of the
application startup sequence.

In this CL, the UNNotificationCenter Delegate initialization was
detached from the Chime registration process and moved up in the
startup sequence.

Bug: 1434079
Change-Id: I36efd7250108be3d9bdcf94b0c3800cdb5a3e273
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4447521
Commit-Queue: Daniel White <danieltwhite@google.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1133888}
NOKEYCHECK=True
GitOrigin-RevId: 92ae6dbe4d1d6c63a268bda7e2562cb80dca2fcd