[iOS] Fix browser check
When removing direct access to BrowserState in crrev.com/c/4418230,
the nullability check on the BrowserState should now be done on the
Browser (as a Browser necessarily have a non-null BrowserState).

Currently, there are some crashes when the Browser is nil and the
code tries to access the BrowserState.

Fixed: 1435126
Change-Id: I41f46662cb397094579f8198756d78cbf842bf9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4458860
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1133948}
NOKEYCHECK=True
GitOrigin-RevId: 03dd2ca4f50169ba61c4dad031ec8fe1a8480f89